feat: Add offset mapping and inference model
Previously the model output needed to be piped through a argmax manually
to get the labels. This was remedied through the extension of the Model
class to an Inference-Model that does this automagically.
The output was further hard to interpret. Thus I added added the offsets
to the batch and wrote a 'pretty-printer'. (I do not enforce the BIO
strictly)