fix: Correct data_manipulation
There is a bug in the original data_manipulation.py which results in
multiple instances of the same training example being in the ..._raw
data when there are multiple annotators for one text, where they are all
copies of the last annotator.
I also moved my toying and thus this fix into a separate subfolder.